+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.1/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.2] - 2026-06-15
 
 ### Added
 
@@ -133,7 +133,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Jinja2 template variables and filters (`sanitize`, `sanitize_db`, `hash_port`)
 - Environment variable overrides for config values
 
-[Unreleased]: https://github.com/tjarvstrand/dojjo/compare/v0.2.1+1...HEAD
+[0.2.2]: https://github.com/tjarvstrand/dojjo/releases/tag/v0.2.2
 [0.2.1+1]: https://github.com/tjarvstrand/dojjo/releases/tag/v0.2.1+1
 [0.2.1]: https://github.com/tjarvstrand/dojjo/releases/tag/v0.2.1
 [0.2.0]: https://github.com/tjarvstrand/dojjo/releases/tag/v0.2.0